+++ src/components/Introduction/components/index.js
@@ -1,0 +1,2 @@
+export { default as IntroDetail } from './introDetail';
+export { default as Sidebar } from './sidebar';